+++ TeamProject/static/js/board/event.js
@@ -1,6 +1,7 @@
 import * as MAIN from "./main.js"
 import * as REND from "./render.js"
 import * as FETCH from "./fetch.js"
+import * as REND_LIST from "./list/render.js";
 //===========보드 메인 포스트 페이지 ==========
 
 //메인화면 페이지로 가는 함수
@@ -178,7 +179,7 @@ export const handle_scrollHeight = async () => {
   if ((window.innerHeight + window.scrollY + footer_size) >= document.body.offsetHeight) {
     SCROLLFLAG = true; //이벤트함수에 접근하고 바로 플래그를 닫는다
     console.log("바닥");
-    REND.infinity_scroll_image(); //로딩창 그려주기
+    REND_LIST.infinity_scroll_image(); //로딩창 그려주기
     //0.5초뒤에 새로운 게시글들을 불러오고 ,그뒤에 플래그를 다시 연다
     setTimeout(() => {
       console.log('0.5초뒤');
